refactor ui kit
