+++ src/app.routes.ts
@@ -1,9 +1,9 @@
 import { Routes } from '@angular/router';
 import { LayoutComponent } from './app/layout/component/layout/layout.component';
 import { DashboardComponent } from './app/pages/dashboard/dashboard/dashboard.component';
-import { Landing } from './app/pages/landing/landing';
 import { NotFoundComponent } from './app/pages/not-found/not-found.component';
 import { DocumentationComponent } from './app/pages/documentation/documentation.component';
+import { LandingComponent } from './app/pages/landing/landing/landing.component';
 
 export const appRoutes: Routes = [
     {
@@ -16,7 +16,7 @@ export const appRoutes: Routes = [
             { path: 'pages', loadChildren: () => import('./app/pages/pages.routes') }
         ]
     },
-    { path: 'landing', component: Landing },
+    { path: 'landing', component: LandingComponent },
     { path: 'notfound', component: NotFoundComponent },
     { path: 'auth', loadChildren: () => import('./app/pages/auth/auth.routes') },
     { path: '**', redirectTo: '/notfound' }
