+++ src/app/pages/uikit/file-demo/file-demo.component.ts
@@ -1,0 +1,31 @@
+import { CommonModule } from '@angular/common';
+import { Component } from '@angular/core';
+import { MessageService } from 'primeng/api';
+import { ButtonModule } from 'primeng/button';
+import { FileUploadModule } from 'primeng/fileupload';
+import { ToastModule } from 'primeng/toast';
+
+@Component({
+    selector: 'app-file-demo',
+    imports: [CommonModule, FileUploadModule, ToastModule, ButtonModule],
+    templateUrl: './file-demo.component.html',
+    styleUrl: './file-demo.component.scss',
+    providers: [MessageService]
+})
+export class FileDemoComponent {
+    uploadedFiles: any[] = [];
+
+    constructor(private messageService: MessageService) {}
+
+    onUpload(event: any) {
+        for (const file of event.files) {
+            this.uploadedFiles.push(file);
+        }
+
+        this.messageService.add({ severity: 'info', summary: 'Success', detail: 'File Uploaded' });
+    }
+
+    onBasicUpload() {
+        this.messageService.add({ severity: 'info', summary: 'Success', detail: 'File Uploaded with Basic Mode' });
+    }
+}
