+++ src/app/pages/uikit/form-layout-demo/form-layout-demo.component.ts
@@ -1,0 +1,23 @@
+import { Component } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { ButtonModule } from 'primeng/button';
+import { FluidModule } from 'primeng/fluid';
+import { InputTextModule } from 'primeng/inputtext';
+import { SelectModule } from 'primeng/select';
+import { TextareaModule } from 'primeng/textarea';
+
+@Component({
+    selector: 'app-formlayout-demo',
+    imports: [InputTextModule, FluidModule, ButtonModule, SelectModule, FormsModule, TextareaModule],
+    templateUrl: './form-layout-demo.component.html',
+    styleUrl: './form-layout-demo.component.scss'
+})
+export class FormLayoutDemoComponent {
+    dropdownItems = [
+        { name: 'Option 1', code: 'Option 1' },
+        { name: 'Option 2', code: 'Option 2' },
+        { name: 'Option 3', code: 'Option 3' }
+    ];
+
+    dropdownItem = null;
+}
