+++ src/app/pages/uikit/input-demo/input-demo.component.ts
@@ -1,0 +1,168 @@
+import { CommonModule } from '@angular/common';
+import { Component, inject } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { TreeNode } from 'primeng/api';
+import { AutoCompleteModule, AutoCompleteCompleteEvent } from 'primeng/autocomplete';
+import { ButtonModule } from 'primeng/button';
+import { CheckboxModule } from 'primeng/checkbox';
+import { ColorPickerModule } from 'primeng/colorpicker';
+import { DatePickerModule } from 'primeng/datepicker';
+import { FloatLabelModule } from 'primeng/floatlabel';
+import { FluidModule } from 'primeng/fluid';
+import { IconFieldModule } from 'primeng/iconfield';
+import { InputGroupModule } from 'primeng/inputgroup';
+import { InputGroupAddonModule } from 'primeng/inputgroupaddon';
+import { InputIconModule } from 'primeng/inputicon';
+import { InputNumberModule } from 'primeng/inputnumber';
+import { InputTextModule } from 'primeng/inputtext';
+import { KnobModule } from 'primeng/knob';
+import { ListboxModule } from 'primeng/listbox';
+import { MultiSelectModule } from 'primeng/multiselect';
+import { RadioButtonModule } from 'primeng/radiobutton';
+import { RatingModule } from 'primeng/rating';
+import { SelectModule } from 'primeng/select';
+import { SelectButtonModule } from 'primeng/selectbutton';
+import { SliderModule } from 'primeng/slider';
+import { TextareaModule } from 'primeng/textarea';
+import { ToggleButtonModule } from 'primeng/togglebutton';
+import { ToggleSwitchModule } from 'primeng/toggleswitch';
+import { TreeSelectModule } from 'primeng/treeselect';
+import { CountryService } from '../../service/country.service';
+import { Country } from '../../service/customer.service';
+import { NodeService } from '../../service/node.service';
+
+@Component({
+    selector: 'app-input-demo',
+    imports: [
+        CommonModule,
+        FormsModule,
+        InputTextModule,
+        ButtonModule,
+        CheckboxModule,
+        RadioButtonModule,
+        SelectButtonModule,
+        InputGroupModule,
+        FluidModule,
+        IconFieldModule,
+        InputIconModule,
+        FloatLabelModule,
+        AutoCompleteModule,
+        InputNumberModule,
+        SliderModule,
+        RatingModule,
+        ColorPickerModule,
+        KnobModule,
+        SelectModule,
+        DatePickerModule,
+        ToggleButtonModule,
+        ToggleSwitchModule,
+        TreeSelectModule,
+        MultiSelectModule,
+        ListboxModule,
+        InputGroupAddonModule,
+        TextareaModule
+    ],
+    templateUrl: './input-demo.component.html',
+    styleUrl: './input-demo.component.scss',
+    providers: [CountryService, NodeService]
+})
+export class InputDemoComponent {
+    floatValue: any = null;
+
+    autoValue: any[] | undefined;
+
+    autoFilteredValue: any[] = [];
+
+    selectedAutoValue: any = null;
+
+    calendarValue: any = null;
+
+    inputNumberValue: any = null;
+
+    sliderValue: number = 50;
+
+    ratingValue: any = null;
+
+    colorValue: string = '#1976D2';
+
+    radioValue: any = null;
+
+    checkboxValue: any[] = [];
+
+    switchValue: boolean = false;
+
+    listboxValues: any[] = [
+        { name: 'New York', code: 'NY' },
+        { name: 'Rome', code: 'RM' },
+        { name: 'London', code: 'LDN' },
+        { name: 'Istanbul', code: 'IST' },
+        { name: 'Paris', code: 'PRS' }
+    ];
+
+    listboxValue: any = null;
+
+    dropdownValues = [
+        { name: 'New York', code: 'NY' },
+        { name: 'Rome', code: 'RM' },
+        { name: 'London', code: 'LDN' },
+        { name: 'Istanbul', code: 'IST' },
+        { name: 'Paris', code: 'PRS' }
+    ];
+
+    dropdownValue: any = null;
+
+    multiselectCountries: Country[] = [
+        { name: 'Australia', code: 'AU' },
+        { name: 'Brazil', code: 'BR' },
+        { name: 'China', code: 'CN' },
+        { name: 'Egypt', code: 'EG' },
+        { name: 'France', code: 'FR' },
+        { name: 'Germany', code: 'DE' },
+        { name: 'India', code: 'IN' },
+        { name: 'Japan', code: 'JP' },
+        { name: 'Spain', code: 'ES' },
+        { name: 'United States', code: 'US' }
+    ];
+
+    multiselectSelectedCountries!: Country[];
+
+    toggleValue: boolean = false;
+
+    selectButtonValue: any = null;
+
+    selectButtonValues: any = [{ name: 'Option 1' }, { name: 'Option 2' }, { name: 'Option 3' }];
+
+    knobValue: number = 50;
+
+    inputGroupValue: boolean = false;
+
+    treeSelectNodes!: TreeNode[];
+
+    selectedNode: any = null;
+
+    countryService = inject(CountryService);
+
+    nodeService = inject(NodeService);
+
+    ngOnInit() {
+        this.countryService.getCountries().then((countries) => {
+            this.autoValue = countries;
+        });
+
+        this.nodeService.getFiles().then((data) => (this.treeSelectNodes = data));
+    }
+
+    filterCountry(event: AutoCompleteCompleteEvent) {
+        const filtered: any[] = [];
+        const query = event.query;
+
+        for (let i = 0; i < (this.autoValue as any[]).length; i++) {
+            const country = (this.autoValue as any[])[i];
+            if (country.name.toLowerCase().indexOf(query.toLowerCase()) == 0) {
+                filtered.push(country);
+            }
+        }
+
+        this.autoFilteredValue = filtered;
+    }
+}
