+++ src/app/pages/uikit/list-demo/list-demo.component.scss
@@ -1,0 +1,5 @@
+::ng-deep {
+    .p-orderlist-list-container {
+        width: 100%;
+    }
+}
