+++ src/app/pages/uikit/list-demo/list-demo.component.ts
@@ -1,0 +1,75 @@
+import { CommonModule } from '@angular/common';
+import { Component } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { ButtonModule } from 'primeng/button';
+import { DataViewModule } from 'primeng/dataview';
+import { OrderListModule } from 'primeng/orderlist';
+import { PickListModule } from 'primeng/picklist';
+import { SelectButtonModule } from 'primeng/selectbutton';
+import { TagModule } from 'primeng/tag';
+import { ProductService, Product } from '../../service/product.service';
+
+@Component({
+    selector: 'app-list-demo',
+    imports: [CommonModule, DataViewModule, FormsModule, SelectButtonModule, PickListModule, OrderListModule, TagModule, ButtonModule],
+    templateUrl: './list-demo.component.html',
+    styleUrl: './list-demo.component.scss',
+    providers: [ProductService]
+})
+export class ListDemoComponent {
+    layout: 'list' | 'grid' = 'list';
+
+    options = ['list', 'grid'];
+
+    products: Product[] = [];
+
+    sourceCities: any[] = [];
+
+    targetCities: any[] = [];
+
+    orderCities: any[] = [];
+
+    constructor(private productService: ProductService) {}
+
+    ngOnInit() {
+        this.productService.getProductsSmall().then((data) => (this.products = data.slice(0, 6)));
+
+        this.sourceCities = [
+            { name: 'San Francisco', code: 'SF' },
+            { name: 'London', code: 'LDN' },
+            { name: 'Paris', code: 'PRS' },
+            { name: 'Istanbul', code: 'IST' },
+            { name: 'Berlin', code: 'BRL' },
+            { name: 'Barcelona', code: 'BRC' },
+            { name: 'Rome', code: 'RM' }
+        ];
+
+        this.targetCities = [];
+
+        this.orderCities = [
+            { name: 'San Francisco', code: 'SF' },
+            { name: 'London', code: 'LDN' },
+            { name: 'Paris', code: 'PRS' },
+            { name: 'Istanbul', code: 'IST' },
+            { name: 'Berlin', code: 'BRL' },
+            { name: 'Barcelona', code: 'BRC' },
+            { name: 'Rome', code: 'RM' }
+        ];
+    }
+
+    getSeverity(product: Product) {
+        switch (product.inventoryStatus) {
+            case 'INSTOCK':
+                return 'success';
+
+            case 'LOWSTOCK':
+                return 'warn';
+
+            case 'OUTOFSTOCK':
+                return 'danger';
+
+            default:
+                return 'info';
+        }
+    }
+}
