+++ src/app/pages/uikit/media-demo/media-demo.component.ts
@@ -1,0 +1,87 @@
+import { CommonModule } from '@angular/common';
+import { Component } from '@angular/core';
+import { ButtonModule } from 'primeng/button';
+import { CarouselModule } from 'primeng/carousel';
+import { GalleriaModule } from 'primeng/galleria';
+import { ImageModule } from 'primeng/image';
+import { TagModule } from 'primeng/tag';
+import { PhotoService } from '../../service/photo.service';
+import { ProductService, Product } from '../../service/product.service';
+
+@Component({
+    selector: 'app-media-demo',
+    standalone: true,
+    imports: [CommonModule, CarouselModule, ButtonModule, GalleriaModule, ImageModule, TagModule],
+    template: ``,
+    providers: [ProductService, PhotoService]
+})
+export class MediaDemoComponent {
+    products!: Product[];
+
+    images!: any[];
+
+    galleriaResponsiveOptions: any[] = [
+        {
+            breakpoint: '1024px',
+            numVisible: 5
+        },
+        {
+            breakpoint: '960px',
+            numVisible: 4
+        },
+        {
+            breakpoint: '768px',
+            numVisible: 3
+        },
+        {
+            breakpoint: '560px',
+            numVisible: 1
+        }
+    ];
+
+    carouselResponsiveOptions: any[] = [
+        {
+            breakpoint: '1024px',
+            numVisible: 3,
+            numScroll: 3
+        },
+        {
+            breakpoint: '768px',
+            numVisible: 2,
+            numScroll: 2
+        },
+        {
+            breakpoint: '560px',
+            numVisible: 1,
+            numScroll: 1
+        }
+    ];
+
+    constructor(
+        private productService: ProductService,
+        private photoService: PhotoService
+    ) {}
+
+    ngOnInit() {
+        this.productService.getProductsSmall().then((products) => {
+            this.products = products;
+        });
+
+        this.photoService.getImages().then((images) => {
+            this.images = images;
+        });
+    }
+
+    getSeverity(status: string) {
+        switch (status) {
+            case 'INSTOCK':
+                return 'success';
+            case 'LOWSTOCK':
+                return 'warn';
+            case 'OUTOFSTOCK':
+                return 'danger';
+            default:
+                return 'success';
+        }
+    }
+}
