+++ src/app/pages/uikit/menu-demo/menu-demo.component.ts
@@ -1,0 +1,450 @@
+import { CommonModule } from '@angular/common';
+import { Component } from '@angular/core';
+import { BreadcrumbModule } from 'primeng/breadcrumb';
+import { ButtonModule } from 'primeng/button';
+import { ContextMenuModule } from 'primeng/contextmenu';
+import { IconFieldModule, IconField } from 'primeng/iconfield';
+import { InputIconModule, InputIcon } from 'primeng/inputicon';
+import { InputTextModule } from 'primeng/inputtext';
+import { MegaMenuModule } from 'primeng/megamenu';
+import { MenuModule } from 'primeng/menu';
+import { MenubarModule } from 'primeng/menubar';
+import { PanelMenuModule } from 'primeng/panelmenu';
+import { StepperModule } from 'primeng/stepper';
+import { TabsModule } from 'primeng/tabs';
+import { TieredMenuModule } from 'primeng/tieredmenu';
+
+@Component({
+    selector: 'app-menu-demo',
+    imports: [
+        CommonModule,
+        BreadcrumbModule,
+        TieredMenuModule,
+        IconFieldModule,
+        InputIconModule,
+        MenuModule,
+        ButtonModule,
+        ContextMenuModule,
+        MegaMenuModule,
+        PanelMenuModule,
+        TabsModule,
+        MenubarModule,
+        InputTextModule,
+        TabsModule,
+        StepperModule,
+        TabsModule,
+        IconField,
+        InputIcon
+    ],
+    templateUrl: './menu-demo.component.html',
+    styleUrl: './menu-demo.component.scss'
+})
+export class MenuDemoComponent {
+    nestedMenuItems = [
+        {
+            label: 'Customers',
+            icon: 'pi pi-fw pi-table',
+            items: [
+                {
+                    label: 'New',
+                    icon: 'pi pi-fw pi-user-plus',
+                    items: [
+                        {
+                            label: 'Customer',
+                            icon: 'pi pi-fw pi-plus'
+                        },
+                        {
+                            label: 'Duplicate',
+                            icon: 'pi pi-fw pi-copy'
+                        }
+                    ]
+                },
+                {
+                    label: 'Edit',
+                    icon: 'pi pi-fw pi-user-edit'
+                }
+            ]
+        },
+        {
+            label: 'Orders',
+            icon: 'pi pi-fw pi-shopping-cart',
+            items: [
+                {
+                    label: 'View',
+                    icon: 'pi pi-fw pi-list'
+                },
+                {
+                    label: 'Search',
+                    icon: 'pi pi-fw pi-search'
+                }
+            ]
+        },
+        {
+            label: 'Shipments',
+            icon: 'pi pi-fw pi-envelope',
+            items: [
+                {
+                    label: 'Tracker',
+                    icon: 'pi pi-fw pi-compass'
+                },
+                {
+                    label: 'Map',
+                    icon: 'pi pi-fw pi-map-marker'
+                },
+                {
+                    label: 'Manage',
+                    icon: 'pi pi-fw pi-pencil'
+                }
+            ]
+        },
+        {
+            label: 'Profile',
+            icon: 'pi pi-fw pi-user',
+            items: [
+                {
+                    label: 'Settings',
+                    icon: 'pi pi-fw pi-cog'
+                },
+                {
+                    label: 'Billing',
+                    icon: 'pi pi-fw pi-file'
+                }
+            ]
+        },
+        {
+            label: 'Quit',
+            icon: 'pi pi-fw pi-sign-out'
+        }
+    ];
+    breadcrumbHome = { icon: 'pi pi-home', to: '/' };
+    breadcrumbItems = [{ label: 'Computer' }, { label: 'Notebook' }, { label: 'Accessories' }, { label: 'Backpacks' }, { label: 'Item' }];
+    tieredMenuItems = [
+        {
+            label: 'Customers',
+            icon: 'pi pi-fw pi-table',
+            items: [
+                {
+                    label: 'New',
+                    icon: 'pi pi-fw pi-user-plus',
+                    items: [
+                        {
+                            label: 'Customer',
+                            icon: 'pi pi-fw pi-plus'
+                        },
+                        {
+                            label: 'Duplicate',
+                            icon: 'pi pi-fw pi-copy'
+                        }
+                    ]
+                },
+                {
+                    label: 'Edit',
+                    icon: 'pi pi-fw pi-user-edit'
+                }
+            ]
+        },
+        {
+            label: 'Orders',
+            icon: 'pi pi-fw pi-shopping-cart',
+            items: [
+                {
+                    label: 'View',
+                    icon: 'pi pi-fw pi-list'
+                },
+                {
+                    label: 'Search',
+                    icon: 'pi pi-fw pi-search'
+                }
+            ]
+        },
+        {
+            label: 'Shipments',
+            icon: 'pi pi-fw pi-envelope',
+            items: [
+                {
+                    label: 'Tracker',
+                    icon: 'pi pi-fw pi-compass'
+                },
+                {
+                    label: 'Map',
+                    icon: 'pi pi-fw pi-map-marker'
+                },
+                {
+                    label: 'Manage',
+                    icon: 'pi pi-fw pi-pencil'
+                }
+            ]
+        },
+        {
+            label: 'Profile',
+            icon: 'pi pi-fw pi-user',
+            items: [
+                {
+                    label: 'Settings',
+                    icon: 'pi pi-fw pi-cog'
+                },
+                {
+                    label: 'Billing',
+                    icon: 'pi pi-fw pi-file'
+                }
+            ]
+        },
+        {
+            separator: true
+        },
+        {
+            label: 'Quit',
+            icon: 'pi pi-fw pi-sign-out'
+        }
+    ];
+    overlayMenuItems = [
+        {
+            label: 'Save',
+            icon: 'pi pi-save'
+        },
+        {
+            label: 'Update',
+            icon: 'pi pi-refresh'
+        },
+        {
+            label: 'Delete',
+            icon: 'pi pi-trash'
+        },
+        {
+            separator: true
+        },
+        {
+            label: 'Home',
+            icon: 'pi pi-home'
+        }
+    ];
+    menuItems = [
+        {
+            label: 'Customers',
+            items: [
+                {
+                    label: 'New',
+                    icon: 'pi pi-fw pi-plus'
+                },
+                {
+                    label: 'Edit',
+                    icon: 'pi pi-fw pi-user-edit'
+                }
+            ]
+        },
+        {
+            label: 'Orders',
+            items: [
+                {
+                    label: 'View',
+                    icon: 'pi pi-fw pi-list'
+                },
+                {
+                    label: 'Search',
+                    icon: 'pi pi-fw pi-search'
+                }
+            ]
+        }
+    ];
+    contextMenuItems = [
+        {
+            label: 'Save',
+            icon: 'pi pi-save'
+        },
+        {
+            label: 'Update',
+            icon: 'pi pi-refresh'
+        },
+        {
+            label: 'Delete',
+            icon: 'pi pi-trash'
+        },
+        {
+            separator: true
+        },
+        {
+            label: 'Options',
+            icon: 'pi pi-cog'
+        }
+    ];
+    megaMenuItems = [
+        {
+            label: 'Fashion',
+            icon: 'pi pi-fw pi-tag',
+            items: [
+                [
+                    {
+                        label: 'Woman',
+                        items: [{ label: 'Woman Item' }, { label: 'Woman Item' }, { label: 'Woman Item' }]
+                    },
+                    {
+                        label: 'Men',
+                        items: [{ label: 'Men Item' }, { label: 'Men Item' }, { label: 'Men Item' }]
+                    }
+                ],
+                [
+                    {
+                        label: 'Kids',
+                        items: [{ label: 'Kids Item' }, { label: 'Kids Item' }]
+                    },
+                    {
+                        label: 'Luggage',
+                        items: [{ label: 'Luggage Item' }, { label: 'Luggage Item' }, { label: 'Luggage Item' }]
+                    }
+                ]
+            ]
+        },
+        {
+            label: 'Electronics',
+            icon: 'pi pi-fw pi-desktop',
+            items: [
+                [
+                    {
+                        label: 'Computer',
+                        items: [{ label: 'Computer Item' }, { label: 'Computer Item' }]
+                    },
+                    {
+                        label: 'Camcorder',
+                        items: [{ label: 'Camcorder Item' }, { label: 'Camcorder Item' }, { label: 'Camcorder Item' }]
+                    }
+                ],
+                [
+                    {
+                        label: 'TV',
+                        items: [{ label: 'TV Item' }, { label: 'TV Item' }]
+                    },
+                    {
+                        label: 'Audio',
+                        items: [{ label: 'Audio Item' }, { label: 'Audio Item' }, { label: 'Audio Item' }]
+                    }
+                ],
+                [
+                    {
+                        label: 'Sports.7',
+                        items: [{ label: 'Sports.7.1' }, { label: 'Sports.7.2' }]
+                    }
+                ]
+            ]
+        },
+        {
+            label: 'Furniture',
+            icon: 'pi pi-fw pi-image',
+            items: [
+                [
+                    {
+                        label: 'Living Room',
+                        items: [{ label: 'Living Room Item' }, { label: 'Living Room Item' }]
+                    },
+                    {
+                        label: 'Kitchen',
+                        items: [{ label: 'Kitchen Item' }, { label: 'Kitchen Item' }, { label: 'Kitchen Item' }]
+                    }
+                ],
+                [
+                    {
+                        label: 'Bedroom',
+                        items: [{ label: 'Bedroom Item' }, { label: 'Bedroom Item' }]
+                    },
+                    {
+                        label: 'Outdoor',
+                        items: [{ label: 'Outdoor Item' }, { label: 'Outdoor Item' }, { label: 'Outdoor Item' }]
+                    }
+                ]
+            ]
+        },
+        {
+            label: 'Sports',
+            icon: 'pi pi-fw pi-star',
+            items: [
+                [
+                    {
+                        label: 'Basketball',
+                        items: [{ label: 'Basketball Item' }, { label: 'Basketball Item' }]
+                    },
+                    {
+                        label: 'Football',
+                        items: [{ label: 'Football Item' }, { label: 'Football Item' }, { label: 'Football Item' }]
+                    }
+                ],
+                [
+                    {
+                        label: 'Tennis',
+                        items: [{ label: 'Tennis Item' }, { label: 'Tennis Item' }]
+                    }
+                ]
+            ]
+        }
+    ];
+    panelMenuItems = [
+        {
+            label: 'Customers',
+            icon: 'pi pi-fw pi-table',
+            items: [
+                {
+                    label: 'New',
+                    icon: 'pi pi-fw pi-user-plus',
+                    items: [
+                        {
+                            label: 'Customer',
+                            icon: 'pi pi-fw pi-plus'
+                        },
+                        {
+                            label: 'Duplicate',
+                            icon: 'pi pi-fw pi-copy'
+                        }
+                    ]
+                },
+                {
+                    label: 'Edit',
+                    icon: 'pi pi-fw pi-user-edit'
+                }
+            ]
+        },
+        {
+            label: 'Orders',
+            icon: 'pi pi-fw pi-shopping-cart',
+            items: [
+                {
+                    label: 'View',
+                    icon: 'pi pi-fw pi-list'
+                },
+                {
+                    label: 'Search',
+                    icon: 'pi pi-fw pi-search'
+                }
+            ]
+        },
+        {
+            label: 'Shipments',
+            icon: 'pi pi-fw pi-envelope',
+            items: [
+                {
+                    label: 'Tracker',
+                    icon: 'pi pi-fw pi-compass'
+                },
+                {
+                    label: 'Map',
+                    icon: 'pi pi-fw pi-map-marker'
+                },
+                {
+                    label: 'Manage',
+                    icon: 'pi pi-fw pi-pencil'
+                }
+            ]
+        },
+        {
+            label: 'Profile',
+            icon: 'pi pi-fw pi-user',
+            items: [
+                {
+                    label: 'Settings',
+                    icon: 'pi pi-fw pi-cog'
+                },
+                {
+                    label: 'Billing',
+                    icon: 'pi pi-fw pi-file'
+                }
+            ]
+        }
+    ];
+}
