+++ src/app/pages/uikit/misc-demo/misc-demo.component.ts
@@ -1,0 +1,39 @@
+import { CommonModule } from '@angular/common';
+import { Component } from '@angular/core';
+import { AvatarModule } from 'primeng/avatar';
+import { AvatarGroupModule } from 'primeng/avatargroup';
+import { BadgeModule } from 'primeng/badge';
+import { ButtonModule } from 'primeng/button';
+import { ChipModule } from 'primeng/chip';
+import { OverlayBadgeModule } from 'primeng/overlaybadge';
+import { ProgressBarModule } from 'primeng/progressbar';
+import { ScrollPanelModule } from 'primeng/scrollpanel';
+import { ScrollTopModule } from 'primeng/scrolltop';
+import { SkeletonModule } from 'primeng/skeleton';
+import { TagModule } from 'primeng/tag';
+
+@Component({
+    selector: 'app-misc-demo',
+    imports: [CommonModule, ProgressBarModule, BadgeModule, AvatarModule, ScrollPanelModule, TagModule, ChipModule, ButtonModule, SkeletonModule, AvatarGroupModule, ScrollTopModule, OverlayBadgeModule],
+    templateUrl: './misc-demo.component.html',
+    styleUrl: './misc-demo.component.scss'
+})
+export class MiscDemoComponent {
+    value = 0;
+
+    interval: any;
+
+    ngOnInit() {
+        this.interval = setInterval(() => {
+            this.value = this.value + Math.floor(Math.random() * 10) + 1;
+            if (this.value >= 100) {
+                this.value = 100;
+                clearInterval(this.interval);
+            }
+        }, 2000);
+    }
+
+    ngOnDestroy() {
+        clearInterval(this.interval);
+    }
+}
