+++ src/app/pages/uikit/panels-demo/panels-demo.component.ts
@@ -1,0 +1,62 @@
+import { CommonModule } from '@angular/common';
+import { Component } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { AccordionModule } from 'primeng/accordion';
+import { MenuItem } from 'primeng/api';
+import { ButtonModule } from 'primeng/button';
+import { DividerModule } from 'primeng/divider';
+import { FieldsetModule } from 'primeng/fieldset';
+import { IconFieldModule } from 'primeng/iconfield';
+import { InputIconModule } from 'primeng/inputicon';
+import { InputTextModule } from 'primeng/inputtext';
+import { MenuModule } from 'primeng/menu';
+import { PanelModule } from 'primeng/panel';
+import { RippleModule } from 'primeng/ripple';
+import { SplitButtonModule } from 'primeng/splitbutton';
+import { SplitterModule } from 'primeng/splitter';
+import { TabsModule } from 'primeng/tabs';
+import { ToolbarModule } from 'primeng/toolbar';
+
+@Component({
+    selector: 'app-panels-demo',
+    imports: [
+        CommonModule,
+        FormsModule,
+        ToolbarModule,
+        ButtonModule,
+        RippleModule,
+        SplitButtonModule,
+        AccordionModule,
+        FieldsetModule,
+        MenuModule,
+        InputTextModule,
+        DividerModule,
+        SplitterModule,
+        PanelModule,
+        TabsModule,
+        IconFieldModule,
+        InputIconModule
+    ],
+    templateUrl: './panels-demo.component.html',
+    styleUrl: './panels-demo.component.scss'
+})
+export class PanelsDemoComponent {
+    items: MenuItem[] = [
+        {
+            label: 'Save',
+            icon: 'pi pi-check'
+        },
+        {
+            label: 'Update',
+            icon: 'pi pi-upload'
+        },
+        {
+            label: 'Delete',
+            icon: 'pi pi-trash'
+        },
+        {
+            label: 'Home Page',
+            icon: 'pi pi-home'
+        }
+    ];
+}
