+++ src/app/pages/uikit/table-demo/table-demo.component.scss
@@ -1,0 +1,7 @@
+.p-datatable-frozen-tbody {
+    font-weight: bold;
+}
+
+.p-datatable-scrollable .p-frozen-column {
+    font-weight: bold;
+}
