+++ src/app/pages/uikit/timeline-demo/timeline-demo.component.html
@@ -1,0 +1,93 @@
+<div class="grid grid-cols-12 gap-8">
+    <div class="col-span-6">
+        <div class="card">
+            <div class="font-semibold text-xl mb-4">Left Align</div>
+            <p-timeline [value]="events1">
+                <ng-template #content let-event>
+                    {{ event.status }}
+                </ng-template>
+            </p-timeline>
+        </div>
+    </div>
+    <div class="col-span-6">
+        <div class="card">
+            <div class="font-semibold text-xl mb-4">Right Align</div>
+            <p-timeline [value]="events1" align="right">
+                <ng-template #content let-event>
+                    {{ event.status }}
+                </ng-template>
+            </p-timeline>
+        </div>
+    </div>
+    <div class="col-span-6">
+        <div class="card">
+            <div class="font-semibold text-xl mb-4">Alternate Align</div>
+            <p-timeline [value]="events1" align="alternate">
+                <ng-template #content let-event>
+                    {{ event.status }}
+                </ng-template>
+            </p-timeline>
+        </div>
+    </div>
+    <div class="col-span-6">
+        <div class="card">
+            <div class="font-semibold text-xl mb-4">Opposite Content</div>
+            <p-timeline [value]="events1">
+                <ng-template #content let-event>
+                    <small class="p-text-secondary">{{ event.date }}</small>
+                </ng-template>
+                <ng-template #opposite let-event>
+                    {{ event.status }}
+                </ng-template>
+            </p-timeline>
+        </div>
+    </div>
+    <div class="col-span-full">
+        <div class="card">
+            <div class="font-semibold text-xl mb-4">Templating</div>
+            <p-timeline [value]="events1" align="alternate" styleClass="customized-timeline">
+                <ng-template #marker let-event>
+                    <span class="flex w-8 h-8 items-center justify-center text-white rounded-full z-10 shadow-sm" [style]="{ 'background-color': event.color }">
+                        <i [class]="event.icon"></i>
+                    </span>
+                </ng-template>
+                <ng-template #content let-event>
+                    <p-card [header]="event.status" [subheader]="event.date">
+                        <img *ngIf="event.image" [src]="'https://primefaces.org/cdn/primeng/images/demo/product/' + event.image" [alt]="event.name" width="200" class="shadow" />
+                        <p>
+                            Lorem ipsum dolor sit amet, consectetur adipisicing elit. Inventore sed consequuntur error repudiandae numquam deserunt quisquam repellat libero asperiores earum nam nobis, culpa ratione quam perferendis esse, cupiditate
+                            neque quas!
+                        </p>
+                        <p-button label="Read more" [text]="true" />
+                    </p-card>
+                </ng-template>
+            </p-timeline>
+        </div>
+    </div>
+    <div class="col-span-full">
+        <div class="card">
+            <div class="font-semibold text-xl mb-4">Horizontal</div>
+            <div class="font-semibold mb-2">Top Align</div>
+            <p-timeline [value]="events2" layout="horizontal" align="top">
+                <ng-template #content let-event>
+                    {{ event }}
+                </ng-template>
+            </p-timeline>
+
+            <div class="font-semibold mt-4 mb-2">Bottom Align</div>
+            <p-timeline [value]="events2" layout="horizontal" align="bottom">
+                <ng-template #content let-event>
+                    {{ event }}
+                </ng-template>
+            </p-timeline>
+
+            <div class="font-semibold mt-4 mb-2">Alternate Align</div>
+            <p-timeline [value]="events2" layout="horizontal" align="alternate">
+                <ng-template #content let-event>
+                    {{ event }}
+                </ng-template>
+                <ng-template #opposite let-event> &nbsp; </ng-template>
+            </p-timeline>
+        </div>
+    </div>
+</div>
