+++ src/app/pages/uikit/tree-demo/tree-demo.component.html
@@ -1,0 +1,26 @@
+<div class="card">
+    <div class="font-semibold text-xl">Tree</div>
+    <p-tree [value]="treeValue" selectionMode="checkbox" [(selection)]="selectedTreeValue"></p-tree>
+</div>
+
+<div class="card">
+    <div class="font-semibold text-xl mb-4">TreeTable</div>
+    <p-treetable [value]="treeTableValue" [columns]="cols" selectionMode="checkbox" [(selectionKeys)]="selectedTreeTableValue" dataKey="key" [scrollable]="true" [tableStyle]="{ 'min-width': '50rem' }">
+        <ng-template #header let-columns>
+            <tr>
+                <th *ngFor="let col of columns">
+                    {{ col.header }}
+                </th>
+            </tr>
+        </ng-template>
+        <ng-template #body let-rowNode let-rowData="rowData" let-columns="columns">
+            <tr [ttRow]="rowNode" [ttSelectableRow]="rowNode">
+                <td *ngFor="let col of columns; let i = index">
+                    <p-treeTableToggler [rowNode]="rowNode" *ngIf="i === 0" />
+                    <p-treeTableCheckbox [value]="rowNode" *ngIf="i === 0" />
+                    {{ rowData[col.field] }}
+                </td>
+            </tr>
+        </ng-template>
+    </p-treetable>
+</div>
