+++ src/app/pages/uikit/tree-demo/tree-demo.component.ts
@@ -1,0 +1,46 @@
+import { CommonModule } from '@angular/common';
+import { Component, inject } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { TreeNode } from 'primeng/api';
+import { TreeModule } from 'primeng/tree';
+import { TreeTableModule } from 'primeng/treetable';
+import { NodeService } from '../../service/node.service';
+
+@Component({
+    selector: 'app-tree-demo',
+    imports: [CommonModule, FormsModule, TreeModule, TreeTableModule],
+    templateUrl: './tree-demo.component.html',
+    styleUrl: './tree-demo.component.scss',
+    providers: [NodeService]
+})
+export class TreeDemoComponent {
+    treeValue: TreeNode[] = [];
+
+    treeTableValue: TreeNode[] = [];
+
+    selectedTreeValue: TreeNode[] = [];
+
+    selectedTreeTableValue = {};
+
+    cols: any[] = [];
+
+    nodeService = inject(NodeService);
+
+    ngOnInit() {
+        this.nodeService.getFiles().then((files) => (this.treeValue = files));
+        this.nodeService.getTreeTableNodes().then((files: any) => (this.treeTableValue = files));
+
+        this.cols = [
+            { field: 'name', header: 'Name' },
+            { field: 'size', header: 'Size' },
+            { field: 'type', header: 'Type' }
+        ];
+
+        this.selectedTreeTableValue = {
+            '0-0': {
+                partialChecked: false,
+                checked: true
+            }
+        };
+    }
+}
